init service

+++ src/MyDyb.Handler/Kafka/Models/Segment.cs
@@ -3,9 +3,9 @@
 public class Segment
 {
     public string Id { get; init; } = default!;
-    public long StartMs { get; init; }
-    public long EndMs { get; init; }
-    public long Speaker { get; init; }
+    public int StartMs { get; init; }
+    public int EndMs { get; init; }
+    public int Speaker { get; init; }
     public string Transcribe { get; init; } = string.Empty;
     public string TranslationRu { get; init; } = string.Empty;
     public string AccentRu { get; init; } = string.Empty;
